requirements.txt
# 实盘交易系统依赖包
# 基于回测验证的SharpeOptimizedStrategy策略

# 核心数据处理
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 技术分析
ta-lib>=0.4.0
yfinance>=0.1.70

# 交易所API
python-binance>=1.0.15
ccxt>=1.60.0

# 机器学习
scikit-learn>=1.0.0
xgboost>=1.5.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# 系统监控
psutil>=5.8.0

# 通知系统
python-telegram-bot>=13.0.0
requests>=2.25.0

# 日志和配置
python-dotenv>=0.19.0
PyYAML>=5.4.0

# 数据库（可选）
sqlalchemy>=1.4.0
sqlite3

# 时间处理
python-dateutil>=2.8.0
pytz>=2021.1

# 异步处理
asyncio
aiohttp>=3.8.0

# 数学计算
numba>=0.56.0

# 测试框架
pytest>=6.0.0
pytest-asyncio>=0.15.0

# 代码质量
black>=21.0.0
flake8>=3.9.0
mypy>=0.910

# 性能分析
memory-profiler>=0.60.0
line-profiler>=3.3.0

# 安全
cryptography>=3.4.0

# 网络
urllib3>=1.26.0
certifi>=2021.5.0

# 其他工具
tqdm>=4.62.0
joblib>=1.1.0
